+++ src/js/videojs.record.js
@@ -202,6 +202,11 @@ class Record extends Plugin {
         this.audioMimeType = recordOptions.audioMimeType;
         this.audioBufferUpdate = recordOptions.audioBufferUpdate;
 
+        // image settings
+        this.imageOutputType = recordOptions.imageOutputType;
+        this.imageOutputFormat = recordOptions.imageOutputFormat;
+        this.imageOutputQuality = recordOptions.imageOutputQuality;
+
         // animation settings
         this.animationFrameRate = recordOptions.animationFrameRate;
         this.animationQuality = recordOptions.animationQuality;
@@ -1442,18 +1447,38 @@ class Record extends Plugin {
      */
     createSnapshot() {
         this.captureFrame().then((result) => {
-            // turn the canvas data into base64 data with a PNG header
-            this.player.recordedData = result.toDataURL('image/png');
+            if (this.imageOutputType === 'blob') {
+                // turn the canvas data into a Blob
+                result.toBlob((blob) => {
+                    this.player.recordedData = blob;
 
-            // hide preview video
-            this.mediaElement.style.display = 'none';
+                    // display the snapshot
+                    this.displaySnapshot();
+                });
+            } else if (this.imageOutputType === 'dataURL') {
+                // turn the canvas data into base64 data
+                this.player.recordedData = result.toDataURL(
+                    this.imageOutputFormat, this.imageOutputQuality);
 
-            // show the snapshot
-            this.player.recordCanvas.show();
+                // display the snapshot
+                this.displaySnapshot();
+            }
+        }, this.imageOutputFormat, this.imageOutputQuality);
+    }
 
-            // stop recording
-            this.stop();
-        });
+    /**
+     * Display snapshot image.
+     * @private
+     */
+    displaySnapshot() {
+        // hide preview video
+        this.mediaElement.style.display = 'none';
+
+        // show the snapshot
+        this.player.recordCanvas.show();
+
+        // stop recording
+        this.stop();
     }
 
     /**